+++ src/commands/handlers/removeRole.ts
@@ -1,7 +1,7 @@
 import { CacheType, ChatInputCommandInteraction, Client } from 'discord.js';
 import { AppDataSource } from '../../typeorm';
 import { SavedRole } from '../../typeorm/entities/SavedRole';
-import { lang } from '../../utils';
+import { lang, logger } from '../../utils';
 
 const tl = lang.removeRole;
 const savedRoleRepo = AppDataSource.getRepository(SavedRole);
@@ -81,7 +81,7 @@ export const removeRoleHandler = async(client: Client, interaction: ChatInputCom
         }
 
     } catch (error) {
-        console.log(tl.fail, error);
+        logger(tl.fail + error, 'ERROR');
         await interaction.reply({
             content: tl.fail,
             ephemeral: true