+++ src/typeorm/index.ts
@@ -1,17 +1,17 @@
 import dotenv from 'dotenv';
 import { DataSource } from 'typeorm';
-import { ArchivedTicket } from './entities/ArchivedTicket';
-import { ArchivedTicketConfig } from './entities/ArchivedTicketConfig';
-import { BotConfig } from './entities/BotConfig';
-import { SavedRole } from './entities/SavedRole';
-import { ServerConfig } from './entities/ServerConfig';
-import { Ticket } from './entities/Ticket';
-import { TicketConfig } from './entities/TicketConfig';
 import { Application } from './entities/application/Application';
 import { ApplicationConfig } from './entities/application/ApplicationConfig';
 import { ArchivedApplication } from './entities/application/ArchivedApplication';
 import { ArchivedApplicationConfig } from './entities/application/ArchivedApplicationConfig';
 import { Position } from './entities/application/Position';
+import { BotConfig } from './entities/BotConfig';
+import { SavedRole } from './entities/SavedRole';
+import { ServerConfig } from './entities/ServerConfig';
+import { ArchivedTicket } from './entities/ticket/ArchivedTicket';
+import { ArchivedTicketConfig } from './entities/ticket/ArchivedTicketConfig';
+import { Ticket } from './entities/ticket/Ticket';
+import { TicketConfig } from './entities/ticket/TicketConfig';
 dotenv.config();
 
 export const AppDataSource = new DataSource({
@@ -23,5 +23,4 @@ export const AppDataSource = new DataSource({
     database: process.env.MYSQL_DB_DATABASE,
     synchronize: true,
     entities: [TicketConfig, Ticket, ArchivedTicketConfig, ArchivedTicket, ServerConfig, SavedRole, BotConfig, Application, ApplicationConfig, ArchivedApplication, ArchivedApplicationConfig, Position],
-    //entities: [TicketConfig, Ticket, ArchivedTicketConfig, ArchivedTicket, ServerConfig, SavedRole, BotConfig, Card, CardAbility, BattleAbilityUses, BattleCardStates, GameSession, PlayerDeck, PlayerStats],
 });